Add test case for live evaluation
