+++ browser/src/UI/components/LiveEvalMarker.less
@@ -17,7 +17,7 @@
     overflow-y: auto;
 
     &.error {
-        border-left: 1px solid red;
+        border-left: 1px dashed red;
     }
 
     .text {
