--- vim/core/oni-plugin-typescript/test/InitialTest.ts
@@ -1,7 +1,0 @@
-import * as assert from "assert"
-
-describe("Initial Test", () => {
-    it("test", () => {
-        assert.ok(true)
-    })
-})
